move cellTItles into facet layout
